jit: (llvm) rename ctype_context to c_type_context
